Add files via upload
Background:
Viscosity (η) measures the fluid’s resistance to flow . It is a very fundamental property of materials as it influences almost every physical property such as softening and melting  etc. Our group used the potential energy approach to find the viscosity at low stresses . 
Very small stresses give rise to very small strain rates as well. As can be seen in Figure 13 the plot for the logarithmic viscosity scales linearly with 𝜎2 at very low stresses and diverges at large stresses35. Low strain rates are very hard to measure in an MD simulation. That’s because the change in the simulation box is very minute and hard to quantify. However, after shearing as shown in schematic Figure 14 every atom breaks bonds and forms other bonds and/or maintains previous bonds. The total number of events occurring at the atomic level is the sum of bond breaking and bond forming events. Since the overall strain has to be a total sum of those individual events, the aim is to be able to relate the total events to the strain rate/time as shown in schematic Figure 15. 
Code description: 
To achieve this the Python code below performs two tasks: 
1.	It would generate the species dependent radial distribution function for the system at every frame so that the total number of nearest neighbors for every atom at every frame dumped can be discerned. 
2.	It would track every atom throughout the duration of the simulation at every frame and calculates the new total number of neighbors as the atom moves to a new location and compares it to the initial neighbors ie at frame 0. It then calculates the total new events for the entire system. 
I plot these two tasks as shown in Figure 16 and Figure 17. Once I gather all this data and calibrate the system I test it at different dumps. I can then scale the events with the strain rate and use that to calculate the viscosity at very small strain rates that would otherwise cannot be quantified accurately.